+++ source/isaaclab_assets/isaaclab_assets/cart_pendulum/mdp/__init__.py
@@ -7,3 +7,4 @@
 
 from .commands import *  # noqa: F403
 from .rewards import *  # noqa: F403
+from .terminations import *  # noqa: F403